Move reminder views to rmndin.users.views
